Implement new cover image conversion optimization
Optimize the image conversion process, since it is the slowest part of
the whole script, by doing the conversion and saving a local copy of the
image in the converted form. This way, if you re-run the script with the
same image, it will use the previously-coverted image rather than
converting it again.

Also did a few other misc. changes:
- fixed up bulletin_find_and_replace.py file in a few other areas
- updated .odt what to work on next doc
- etc