wip: working on prettifying the xml in extracted .odt/.ods docs
Currently the "content.xml" files are literally 1 single line that is
tens of thousands of characters. This is impossible to `git diff` or
`git difftool`, so I'm going to "prettify" the xml by running an auto-
formatter on it to make it easy to read, and properly formatted on many
lines instead. Then, it will be `git difftool`able.